📝 Chore(#228): create-steady 상수 파일 변경
- labelData, selectorItem 두 파일로 분리
- import 경로 수정
