+++ src/constants/labelData.ts
@@ -1,0 +1,18 @@
+export const STEADY_RECRUITMENT_EXAMPLE =
+  "[모집 내용 예시]\n" +
+  "∙ 주제 : 프론트엔드 스터디\n" +
+  "∙ 목표 : 프론트엔드 기초 공부\n" +
+  "∙ 예상 일정(횟수) : 2주(2회)\n" +
+  "∙ 예상 커리큘럼 간략히 : HTML, CSS, JavaScript\n" +
+  "∙ 예상 모집인원 : 4명\n" +
+  "∙ 소개와 개설 이유 : 프론트엔드에 입문하고 싶은 사람끼리 모여서 함께 공부하고 싶어서 개설했습니다.\n" +
+  "∙ 주의사항 : 2주에 2회 이상 참여하지 못하면 강퇴됩니다.\n" +
+  "∙ 연락할 수 있는 방법을 남겨주세요. (이메일, 카카오 오픈채팅방 등...) : 카카오톡 오픈채팅방 링크\n";
+
+export const CREATE_STEADY_PAGE_HEADING =
+  "📖 스테디 / ✍️ 모집글 정보를 입력해주세요.";
+
+export const steadyCategoriesWithEmoji = {
+  STUDY: "📖 스터디",
+  PROJECT: "🖥️ 프로젝트",
+};
